更新项目接口

+++ src/main/java/com/sw/chefubao/vo/OrderVo.java
@@ -18,7 +18,7 @@ public class OrderVo {
     private Integer id;
     private String orderNo;
     private Integer orderReceiverAddressId;
-    private Integer payTotal;
+    private Double payTotal;
     private Integer status;
     private List<OrderProductVo> products;
